Format, MIDI, Joystic
Fortmat command get stuck after the "Send Command" part. it's returning error buffer underrun.

Add code for the Analog Joystic : need to remode the ferit to not disturb the 480k or 1M resistance with the 74LV245's pulupp

Add also the code for the MIDI loop back usind the UART mode of the MPU-401